+++ src/onegov.org/onegov/org/upgrade.py
@@ -3,7 +3,10 @@ upgraded on the server. See :class:`onegov.core.upgrade.upgrade_task`.
 
 """
 from onegov.core.upgrade import upgrade_task
-from onegov.org.models import Organisation
+from onegov.form import FormDefinition
+from onegov.libres.models import Resource
+from onegov.org.models import Organisation, Topic, News
+from onegov.org.utils import annotate_html
 
 
 @upgrade_task('Move from town to organisation', always_run=True)
@@ -31,3 +34,18 @@ def move_town_to_organisation(context):
 
     context.operations.drop_table('organisations')
     context.operations.rename_table('towns', 'organisations')
+
+
+@upgrade_task('Add image-dimensions to html')
+def add_image_dimensions_to_html(context):
+    for topic in context.session.query(Topic):
+        topic.text = annotate_html(topic.text, context.request)
+
+    for news in context.session.query(News):
+        news.text = annotate_html(news.text, context.request)
+
+    for resource in context.session.query(Resource):
+        resource.text = annotate_html(resource.text, context.request)
+
+    for form in context.session.query(FormDefinition):
+        form.text = annotate_html(form.text, context.request)